fixed fbid input

+++ migrations/20160303130350_create_users.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
     table.increments();
     table.string('name');
     table.string('profilepicture');
-    table.integer('fb_id');
+    table.float('fb_id');
   })
 };
